requirements.txt
streamlit
tensorflow
keras
numpy
pillow
opencv-python
pandas
plotly
scikit-learn
